Reworked Parser, moved Image to Texture

+++ src/Filters/Blur.cpp
@@ -1,6 +1,6 @@
 #include "Blur.h"
 
-Blur::Blur(size_t radius)
+Blur::Blur(float radius)
 : radius(radius)
 {
 }
