+++ src/Filters/Blur.h
@@ -7,10 +7,10 @@ class Blur : public Filter
 {
 
 	private:
-		size_t radius;
+		float radius;
 
 	public:
-		Blur(size_t radius);
+		Blur(float radius);
 		Vec4 operator()(Vec4 *colorBuffer, float *zBuffer, size_t x, size_t y, size_t width, size_t height);
 
 };
